Added some new tables and cors cannot pass now, finding reason

+++ backend/src/main/java/com/tinysnow/system/model/Tag.java
@@ -3,8 +3,9 @@ package com.tinysnow.system.model;
 import io.mybatis.provider.Entity;
 import lombok.Data;
 
+
 /**
- * tag -
+ * tag - 
  *
  * @author Snow
  */
@@ -12,7 +13,7 @@ import lombok.Data;
 @Entity.Table(value = "tag", remark = "", autoResultMap = true)
 public class Tag {
   @Entity.Column(value = "id", id = true, remark = "标签 id", updatable = false, insertable = false)
-  private Integer id;
+  private Long id;
 
   @Entity.Column(value = "label", remark = "标签名称")
   private String label;
